feat: valid method deleteTask

+++ src/task.js
@@ -4,32 +4,32 @@ class Task {
   static async ToDoList(taskData) {
     const task = await taskData;
     const validTitle = this.validTitle(task)
-    if(!validTitle.valid) throw new Error(validTitle.error)
+    if (!validTitle.valid) throw new Error(validTitle.error)
     const validDescription = this.validDescription(task)
-    if(!validDescription.valid) throw new Error(validDescription.error)
+    if (!validDescription.valid) throw new Error(validDescription.error)
     const validDueDate = this.validDate(task)
-    if(!validDueDate.valid) throw new Error(validDueDate.error)
+    if (!validDueDate.valid) throw new Error(validDueDate.error)
     return this.insertTask(task)
   }
 
   static validTitle(taskData) {
     const title = taskData.title;
 
-    if(!title){
+    if (!title) {
       return {
         error: error.taskTitle.required,
         valid: false,
       }
     }
 
-    if(title.length <= 3) {
+    if (title.length <= 3) {
       return {
         error: error.taskTitle.tooShort,
         valid: false,
       }
     }
 
-    if(title.length >= 100) {
+    if (title.length >= 100) {
       return {
         error: error.taskTitle.tooLong,
         valid: false,
@@ -37,7 +37,7 @@ class Task {
     }
 
     const containsSpecialCharacters = /[^a-zA-Z0-9 ]/.test(title);
-    if(containsSpecialCharacters) {
+    if (containsSpecialCharacters) {
       return {
         error: error.taskTitle.invalidCharacter,
         valid: false,
@@ -48,21 +48,21 @@ class Task {
 
   static validDescription(taskData) {
     const description = taskData.description
-    if(!description){
+    if (!description) {
       return {
         error: error.description.required,
         valid: false,
       }
     }
 
-    if(description.length <= 10) {
+    if (description.length <= 10) {
       return {
         error: error.description.tooShort,
         valid: false,
       }
     }
 
-    if(description.length >= 500) {
+    if (description.length >= 500) {
       return {
         error: error.description.tooLong,
         valid: false,
@@ -80,7 +80,7 @@ class Task {
         valid: false,
       };
     }
-  
+
     const regex = /^\d{4}-\d{2}-\d{2}$/;
     if (!regex.test(dueDate)) {
       return {
@@ -88,7 +88,7 @@ class Task {
         valid: false,
       };
     }
-  
+
     const date = new Date(dueDate);
     const timestamp = date.getTime();
     if (Number.isNaN(timestamp)) {
@@ -97,14 +97,14 @@ class Task {
         valid: false,
       };
     }
-  
+
     if (new Date(dueDate) < new Date(creationDate)) {
       return {
         error: error.dueDate.pastDate,
         valid: false,
       };
     }
-  
+
     return { valid: true };
   }
 
@@ -120,10 +120,18 @@ class Task {
   }
 
   static async deleteTask(id) {
+    if (!id) return { status: 'error', statusCotode: 400, message: 'Invalid ID' }
+
     const tasks = JSON.parse(localStorage.getItem('tasks'));
+    const taskExists = tasks.some(task => task.id === id);
+
+    if (!taskExists) {
+        return { status: 'error', statusCode: 404, message: 'Task not found' };
+    }
+    
     const newTasks = id ? tasks.filter(task => task.id !== id) : tasks
     localStorage.setItem('tasks', JSON.stringify(newTasks));
-    
+
     return {
       status: 'success',
       statusCode: 200,
